+++ apps/frontend/src/libs/enums/notification-massage.enum.ts
@@ -1,5 +1,6 @@
 const NotificationMessage = {
 	AUTH_FAILED: "Authentication failed. Please log in again.",
+	EDIT_PROFILE_SUCCESS: "Profile was edited successfully",
 	SIGN_IN_FAILED:
 		"We couldn't sign you in. Please check your email and password.",
 	SIGN_UP_FAILED: "Something went wrong during sign up. Please try again.",